backend/api/requirements.txt
fastapi==0.115.8
pydantic==2.10.6
pydantic-settings==2.7.1
pynacl==1.5.0
pyjwt==2.10.1
uvicorn==0.34.0

